chore: better file restructuring
