+++ apps/app-sandbox-rnative/src/app/_layout.tsx
@@ -6,16 +6,13 @@ import {
   GlobalTooltipBottomSheet,
   ThemeProvider,
 } from '@ledgerhq/lumen-ui-rnative';
-import { useTheme } from '@ledgerhq/lumen-ui-rnative/styles';
 import { Stack } from 'expo-router';
 import { useState } from 'react';
 import type { ColorSchemeName } from 'react-native';
-import { StatusBar } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import { SafeAreaProvider } from 'react-native-safe-area-context';
-import { SandboxProvider } from '../components/SandboxContext';
-import { SandboxHeader } from '../components/SandboxHeader';
-import { ThemeControlsProvider } from '../components/ThemeControls';
+import { SandboxHeader } from '../components/ui/Header';
+import { SandboxProvider } from '../contexts/SandboxContext';
+import { ThemeControlsProvider } from '../hooks/useThemeControls';
 
 export default function RootLayout() {
   const [colorScheme, setColorScheme] = useState<ColorSchemeName>('dark');
@@ -23,39 +20,29 @@ export default function RootLayout() {
 
   return (
     <GestureHandlerRootView style={{ flex: 1 }}>
-      <SafeAreaProvider>
-        <ThemeProvider
-          themes={ledgerLiveThemes}
-          colorScheme={colorScheme}
-          locale={locale}
+      <ThemeProvider
+        themes={ledgerLiveThemes}
+        colorScheme={colorScheme}
+        locale={locale}
+      >
+        <ThemeControlsProvider
+          value={{ colorScheme, setColorScheme, locale, setLocale }}
         >
-          <ThemeControlsProvider
-            value={{ colorScheme, setColorScheme, locale, setLocale }}
-          >
-            <SandboxProvider>
-              <BottomSheetModalProvider>
-                <StatusBar />
-                <RootNavigator />
-                <GlobalTooltipBottomSheet />
-              </BottomSheetModalProvider>
-            </SandboxProvider>
-          </ThemeControlsProvider>
-        </ThemeProvider>
-      </SafeAreaProvider>
+          <SandboxProvider>
+            <BottomSheetModalProvider>
+              <RootNavigator />
+              <GlobalTooltipBottomSheet />
+            </BottomSheetModalProvider>
+          </SandboxProvider>
+        </ThemeControlsProvider>
+      </ThemeProvider>
     </GestureHandlerRootView>
   );
 }
 
 const RootNavigator = () => {
-  const { theme } = useTheme();
-
   return (
-    <Stack
-      screenOptions={{
-        header: () => <SandboxHeader />,
-        contentStyle: { backgroundColor: theme.colors.bg.canvas },
-      }}
-    >
+    <Stack screenOptions={{ header: () => <SandboxHeader /> }}>
       <Stack.Screen name='index' />
     </Stack>
   );
